+++ public/admin/_app/immutable/nodes/14.Dj2kiT1R.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/B_hRuOHx2.js";export{e as component};