+++ public/admin/_app/immutable/nodes/17.CAwPP9Nl.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/BBDnF_6G2.js";export{e as component};