+++ public/admin/_app/immutable/nodes/18.DS5jZQCh.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/DyXI-h_C2.js";export{e as component};