+++ public/admin/_app/immutable/nodes/20.-sfx9NK4.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/Ckwx16Ro.js";export{e as component};