+++ public/admin/_app/immutable/nodes/23.CNePyQg_.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/DoMbU5oX.js";export{e as component};