+++ public/admin/_app/immutable/nodes/24.BxIIBBFf.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/C-WXw9Zk.js";export{e as component};